+++ core/src/files.js
@@ -11,6 +11,7 @@ export * from './memory.js';
 export * from './artifacts.js';
 export * from './packets.js';
 export * from './tasks.js';
+export * from './adapters.js';
 export * from './schema.js';
 export { ulid } from './ids.js';
 export { slugify } from './slug.js';